add net_exchange_gameplay.c
