+++ linux.mk
@@ -44,6 +44,7 @@ src/bflib_mspointer.cpp \
 src/bflib_netsession.c \
 src/bflib_netsp.cpp \
 src/net_exchange_common.c \
+src/net_exchange_gameplay.c \
 src/net_main.c \
 src/net_lobby.c \
 src/bflib_planar.c \
