+++ src/net_exchange_gameplay.h
@@ -1,0 +1,34 @@
+/******************************************************************************/
+// Free implementation of Bullfrog's Dungeon Keeper strategy game.
+/******************************************************************************/
+/** @file net_exchange_gameplay.h
+ *     Header file for net_exchange_gameplay.c.
+ * @par Purpose:
+ *     Gameplay-specific multiplayer exchange routines.
+ */
+/******************************************************************************/
+#ifndef DK_NET_EXCHANGE_GAMEPLAY_H
+#define DK_NET_EXCHANGE_GAMEPLAY_H
+
+#include "bflib_basics.h"
+#include "globals.h"
+#include "net_main.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+struct Packet;
+
+size_t write_gameplay_frame_payload(NetUserId source_id, const void *send_buf, char *write_pos);
+void send_gameplay_frame_to_peer(NetUserId peer_id, const char *buffer, size_t message_size);
+void store_sent_gameplay_frame(NetUserId source_id, const void *send_buf);
+TbError process_gameplay_frame_payload(NetUserId peer_id, char *read_pos, void *server_buf, size_t frame_size);
+void LbNetwork_WaitForMissingPackets(void *server_buf, size_t client_frame_size);
+void LbNetwork_BroadcastUnpauseTimesync(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
