+++ src/packets.c
@@ -37,6 +37,7 @@
 #include "net_main.h"
 #include "net_lobby.h"
 #include "net_exchange_common.h"
+#include "net_exchange_gameplay.h"
 #include "bflib_sound.h"
 #include "bflib_sndlib.h"
 #include "bflib_sprfnt.h"
